Struct secret_service::SecretService

pub struct SecretService {
    // some fields omitted
}

Secret Service Struct.

This the main entry point for usage of the library.

Creating a new SecretService will also initialize dbus and negotiate a new cryptographic session (EncryptionType::Plain or EncryptionType::Dh)

Methods

impl SecretService

fn new(encryption: EncryptionType) -> Result<Self>

Create a new SecretService instance

Example

let ss = SecretService::new(EncryptionType::Dh).unwrap();

fn get_all_collections(&self) -> Result<Vec<Collection>>

Get all collections

fn get_collection_by_alias(&self, alias: &str) -> Result<Collection>

Get collection by alias. Most common would be the default alias, but there is also a specific method for getting the collection by default aliasl

fn get_default_collection(&self) -> Result<Collection>

Get default collection. (The collection whos alias is default)

fn get_any_collection(&self) -> Result<Collection>

Get any collection. First tries default collection, then session collection, then the first collection when it gets all collections.

fn create_collection(&self, label: &str, alias: &str) -> Result<Collection>

Creates a new collection with a label and an alias.

fn search_items(&self, attributes: Vec<(&str, &str)>) -> Result<Vec<Item>>

Searches all items by attributes